requirements.txt
argon2-cffi==18.1.0
cffi==1.11.5
mysql-connector==2.1.6
pkg-resources==0.0.0
pycparser==2.18
six==1.11.0
